fix: correctly align project card info when some cards have multi-line names
This change updates the styling of the project card bodies so that
alignment of the project data (number of flags, members, health) stays
at the bottom of the project card body, even if some cards have longer
bodies than others.

Before this change, the bottom row would be misaligned if some (but
not all) project cards on a row had names long enough to span multiple lines.