chore: make feature.spec and segements.spec more resilient (#7289)
This PR attempts to make the feature.spec and segements.spec test suites
more reliable. They have been flaking out a lot recently, and this will
hopefully make them less flaky.

The way of handling it is a little different for each test suite.

## feature.spec

Some of the failures we're seeing for the feature/feauture.spec test
suite are due to uncaught resize observer issues (possibly triggered by
the banners).

We can ignore these errors as they don't impact functionality, only
rendering, and are likely to resolve themselves quickly in real-world
scenarios.

On the other hand, it might also ignore actual errors, so I'm not a 100%
on this. Would love some input.

However, MDN has some info on [observation
errors](https://developer.mozilla.org/en-US/docs/Web/API/ResizeObserver#observation_errors):

> As long as the error event does not fire indefinitely, resize observer
will settle and produce a stable, likely correct, layout. However,
visitors may see a flash of broken layout, as a sequence of changes
expected to happen in a single frame is instead happening over multiple
frames.

Based on that, I think this is a pretty safe error to ignore. 

I'm unsure whether catching this exception is only set in the `after`
cleanup or whether it pollutes the cy object for all tests, but I think
it's fine either way. But if you have ideas, I'd love to hear them.

## segments.spec

The issue here appears to be that when we first input the segment's name
in the form, it takes a little time for the UI to become ready, so the
first characters of the string are cut off.

This is a known [issue that the cypress team are
aware](https://github.com/cypress-io/cypress/issues/3817), but that
isn't likely to get fixed any time soon because no one can give them a
reproducible example.

You can see the effect of this on segments that haven't been cleaned up
in the preview:

![image](https://github.com/Unleash/unleash/assets/17786332/1db59906-a2ee-4149-869b-81f2245b4399)

To work around it, we add a 500ms wait before we start filling out the
form. Yes, adding [waits in your tests is an
antipattern](https://docs.cypress.io/guides/references/best-practices#Unnecessary-Waiting),
but it's the easiest way around in this case.

We *could* investigate and find a way not to need that, but that would
likely be a much larger project. This appears to mitigate the issue
immediately, so is at least a pretty good temporary fix in my opinion.
We also already do this in other tests, so there is a precedent for it.