ignore .yarn changes and delete existing tracked files (#7333)
This PR follows Yarn's official suggestion on [which files should be
gitignored](https://yarnpkg.com/getting-started/qa#which-files-should-be-gitignored)
and also deletes the `install-state.gz` files that are already
committed.

I've used the gitignore option that isn't for "zero-installs" because
I'm not sure whether we're using them or not. I'm guessing we're not,
but we can always update that later if we are.

The doc says that the install-state file "is an optimization file that
you shouldn't ever have to commit. It simply stores the exact state of
your project so that the next commands can boot without having to
resolve your workspaces all over again."