chore: wait to input the name of the segment when checking for error messages
This change adds a wait statement before entering the name of the
segment when checking for error messages that this segment name
already exists.

This is the same workaround that we did in
https://github.com/Unleash/unleash/pull/7289, which seems to have
worked.

Like in that PR, using waits is still an antipattern, but it appears
to be working.