docs: document resource limits (#7567)
This PR adds the first version of the resource limit documentation. It
also corrects the link in the `Limit` component.

In addition to the new limits from the soft limits project, I've also
added the limits from the signals project.

## Discussion points

I've set the Pro resource limits to projects and envs to the same value
as on the [plans & pricing](https://www.getunleash.io/pricing) page
instead of the actual hard soft limit. I think that makes messaging
easier and more consistent.

However, this does present us with a bit of a problem. Because we've
been pretty consistent at saying "no, you can't have more projects" for
Pro. But we're saying it's easy to change (when in actuality the limit
isn't enforced 💁🏼 )

## Screenie

It looks a little bit like this (but check the files or preview for the
actual text content)

![image](https://github.com/Unleash/unleash/assets/17786332/56cadc00-5a8e-4820-8b11-d059112934da)